fix(support): attachments not reaching the AI model
Two bugs prevented image attachments from being forwarded to Claude:

1. Double-wrapping in prompt.ts — formatMessageText was calling
   wrapCustomerText(m.body) on bodies that index.ts had already
   wrapped. The double-nested <customer_message> tags and the
   "[Attached images…]" text fallback told Claude an image existed
   but the actual image block was never included.

2. Silent error swallowing in index.ts — the try/catch around the
   file-read block logged nothing when existsSync returned false or
   readFile threw, making the root cause invisible.

Fixes:
- prompt.ts: use m.body directly (already wrapped by the caller)
- storage.ts: add readAttachmentBuffer() — single function handles
  both local files and R2, with warn/error logging when a file is
  missing or unreadable
- index.ts: replace duplicated inline R2/local read logic with
  readAttachmentBuffer(), surface errors via console.error
